将CRC也用ASM加速了

+++ source/scsd/new_scsdio.h
@@ -28,7 +28,7 @@ void send_clk(u32 num);
 void sc_mode(u16 mode);
 void sc_sdcard_reset(void);
 void SDCommand(u8 command,u32 sector);
-void sd_data_write(u16 *buff,u16* crc16buff);
+void sd_data_write(u16 *buff,u8* crc16buff);
 void WriteSector(u8 *buff,u32 sector,u32 writenum);
 void ReadSector (u8 *buff,u32 sector,u32 readnum);
